Fix: unnecessary update triggered when _all_ or _some_ certs are deleted permanently
`RevocationListUpdater` is using a reactive way to trigger the revocation list update and that works for the required cases of registering a new certificate either by scanning
or retrieving it from the back-end. As side effect this re-active way also triggers when a certificate is deleted permanently from the App.

This last point shall be optimised to avoid unnecessary updates. (FIXED by this commit)